servicedb/requirements.txt
fastapi
psycopg2-binary
pika